requirements.txt
asgiref==3.8.1
Django==5.0.6
django-environ==0.11.2
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
mysqlclient==2.2.4
packaging==24.1
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.5.0
typing_extensions==4.12.2
uritemplate==4.1.1
